Move sync to deprecated_ prefix
