+++ beacon-chain/deprecated_sync/BUILD.bazel
@@ -9,14 +9,14 @@ go_library(
         "regular_sync.go",
         "service.go",
     ],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/sync",
+    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/deprecated_sync",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
         "//beacon-chain/blockchain:go_default_library",
         "//beacon-chain/core/helpers:go_default_library",
         "//beacon-chain/db:go_default_library",
+        "//beacon-chain/deprecated_sync/initial-sync:go_default_library",
         "//beacon-chain/operations:go_default_library",
-        "//beacon-chain/sync/initial-sync:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//proto/eth/v1alpha1:go_default_library",
         "//shared/bytesutil:go_default_library",
