+++ beacon-chain/deprecated_sync/initial-sync/helpers.go
