+++ beacon-chain/deprecated_sync/initial-sync/sync_blocks.go
