+++ beacon-chain/deprecated_sync/initial-sync/sync_state.go
