+++ beacon-chain/deprecated_sync/receive_block.go
@@ -1,4 +1,4 @@
-package sync
+package deprecated_sync
 
 import (
 	"context"
