+++ beacon-chain/deprecated_sync/regular_sync.go
@@ -1,5 +1,5 @@
 // Package sync defines the utilities for the beacon-chain to sync with the network.
-package sync
+package deprecated_sync
 
 import (
 	"bytes"
