+++ beacon-chain/node/BUILD.bazel
@@ -13,12 +13,12 @@ go_library(
         "//beacon-chain/attestation:go_default_library",
         "//beacon-chain/blockchain:go_default_library",
         "//beacon-chain/db:go_default_library",
+        "//beacon-chain/deprecated_sync:go_default_library",
         "//beacon-chain/flags:go_default_library",
         "//beacon-chain/gateway:go_default_library",
         "//beacon-chain/operations:go_default_library",
         "//beacon-chain/powchain:go_default_library",
         "//beacon-chain/rpc:go_default_library",
-        "//beacon-chain/sync:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//proto/eth/v1alpha1:go_default_library",
         "//shared:go_default_library",
