+++ beacon-chain/rpc/BUILD.bazel
@@ -22,8 +22,8 @@ go_library(
         "//beacon-chain/core/state:go_default_library",
         "//beacon-chain/core/state/stateutils:go_default_library",
         "//beacon-chain/db:go_default_library",
+        "//beacon-chain/deprecated_sync:go_default_library",
         "//beacon-chain/operations:go_default_library",
-        "//beacon-chain/sync:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//proto/beacon/rpc/v1:go_default_library",
         "//proto/eth/v1alpha1:go_default_library",
